refactor: add LoaderLoadingError class for improved error handling
