+++ lib/errors/LoaderLoadingError.js
@@ -5,7 +5,9 @@
 
 "use strict";
 
-class LoadingLoaderError extends Error {
+const WebpackError = require("./WebpackError");
+
+class LoaderLoadingError extends WebpackError {
 	/**
 	 * @param {string} message message
 	 */
@@ -15,4 +17,4 @@ class LoadingLoaderError extends Error {
 	}
 }
 
-module.exports = LoadingLoaderError;
+module.exports = LoaderLoadingError;
