+++ lib/loaders/loadLoader.js
@@ -5,7 +5,7 @@
 
 "use strict";
 
-const LoaderLoadingError = require("./LoaderLoadingError");
+const LoaderLoadingError = require("../errors/LoaderLoadingError");
 
 /** @typedef {import("./LoaderRunner").LoaderObject} LoaderObject */
